requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==21.4.0
backoff==1.8.0
black==21.12b0
certifi==2021.10.8
chardet==3.0.4
ciso8601==2.2.0
click==8.0.3
idna==2.7
jsonschema==2.6.0
multidict==5.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
python-dateutil==2.8.2
pytz==2018.4
simplejson==3.11.1
singer-python==5.9.0
six==1.16.0
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.24.3
yarl==1.7.2
